// src/app/modules/grid/models/cell-value-type.enum.ts
/**
 * Тип значения в ячейке грида
 */
export enum CellValueType {
    Int16 = 'Int16',
    UInt16 = 'UInt16',
    Int32 = 'Int32',
    UInt32 = 'UInt32',
    Int64 = 'Int64',
    UInt64 = 'UInt64',
    Single = 'Single',
    Decimal = 'Decimal',
    Double = 'Double',
    String = 'String',
    DateTime = 'DateTime',
    Boolean = 'Boolean'
}